python-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-multipart==0.0.12
firebase-admin==6.5.0
google-cloud-firestore==2.18.0
google-cloud-storage==2.18.2
httpx==0.27.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
websockets==12.0
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
joblib==1.4.2
geopy==2.4.1
rsa==4.9
Pillow==10.4.0
qrcode==7.4.2
validators==0.34.0
web3==6.20.1
eth-account==0.13.3
openai==1.45.0
protobuf==4.25.3
